question module is fixed

+++ src/modules/question/question.service.ts
@@ -6,6 +6,8 @@ import { Promise } from 'mongoose';
 import { CreateQuestionDto } from './dto/create-question.dto';
 import { UpdateAboutDto } from '../about/dto/update-about.dto';
 import { UpdateQuestionDto } from './dto/update-question.dto';
+import { QuestionResponse } from './dto/response-question.dto';
+import { DeleteQuestionResponseDto } from './dto/delete-question.dto';
 
 @Injectable()
 export class QuestionService {
@@ -14,39 +16,47 @@ export class QuestionService {
         @InjectModel(Question.name) private questionModel: Model<Question>
     ){}
 
+    private query = "question answer"
 
-    async create (createQuestionDto: CreateQuestionDto): Promise<CreateQuestionDto[]> {
-        const question = new this.questionModel(createQuestionDto);
+
+    async create (userId:string, createQuestionDto: CreateQuestionDto): Promise<QuestionResponse[]> {
+        const question = new this.questionModel({...createQuestionDto, userId: userId});
         await question.save();
-        return await this.questionModel.find({userId: createQuestionDto.userId}).exec()
+        return await this.questionModel.find({userId: userId}).select(this.query).exec()
     }
 
-    async update (id:string, updateQuestionDto: UpdateQuestionDto): Promise<CreateQuestionDto[]> {
-        if(!Types.ObjectId.isValid(id)){
+    async update (id:string, userId: string, updateQuestionDto: UpdateQuestionDto): Promise<QuestionResponse[]> {
+        if(!Types.ObjectId.isValid(id) || !Types.ObjectId.isValid(userId)){
             throw new BadRequestException("Invalid item id")
         }
 
-        const updated = await  this.questionModel.findOneAndUpdate({_id: id, userId:updateQuestionDto.userId}, updateQuestionDto, {new:true, runValidators:true})
+        const updated = await  this.questionModel.findOneAndUpdate({_id: id, userId:userId}, updateQuestionDto, {new:true, runValidators:true})
 
         if(!updated){
             throw new NotFoundException("There is no item");
         }
 
-        return await this.questionModel.find({userId: updateQuestionDto.userId}).exec()
+        return await this.questionModel.find({userId: userId}).select(this.query).exec()
 
     }
 
 
-    async access (userId: string): Promise<CreateQuestionDto[]> {
+    async access (userId: string): Promise<QuestionResponse[]> {
         if(!Types.ObjectId.isValid(userId)){
             throw new BadRequestException("Invalid user id")
         }
-        return await this.questionModel.find({userId: userId}).exec()
+        return await this.questionModel.find({userId: userId}).select(this.query).exec()
     }
 
-    async remove (id:string, userId:string): Promise<any> {
+    async remove (id:string, userId:string): Promise<DeleteQuestionResponseDto> {
         if(!Types.ObjectId.isValid(userId) || !Types.ObjectId.isValid(userId)){
             throw new BadRequestException("Invalid user id or item id")
+        }
+
+        const isFound = await this.questionModel.findOne({_id: id, userId: userId}).select(this.query).exec()
+
+        if(!isFound){
+            throw new NotFoundException("There is no question is found")
         }
 
         const removed = await this.questionModel.deleteOne({_id: id, userId:userId}).exec()
@@ -55,7 +65,12 @@ export class QuestionService {
             throw new NotFoundException("There is no item");
         }
 
-        return removed
+        const remaining = await this.questionModel.find({userId:userId}).select(this.query).exec();
+
+        return {
+            removed: isFound,
+            remaining
+        }
 
     }
